requirements.txt
fastapi
uvicorn
tweepy
transformers
pydantic
streamlit
python-dotenv
